requirement.txt
# tensorflow
numpy
transformers
bert-extractive-summarizer
spacy
googletrans
google-search-results
pandas
Flask
scikit-learn
requests
Pillow
gunicorn == 20.0.4